requirements.txt
ansible
tcconfig
django-notifications-hq
django-extensions

-e  https://github.com/google/ipaddr-py.git#egg=ipaddr
